requirements.txt
Jupyter
Django
djangorestframework